+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Sync scroll in markdown editor and preview
+* Pin ghostscript version to fix bugs in pdf compression
 * UI: Add a button to expand collapsed menus