+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Sync scroll in markdown editor and preview
 * Pin ghostscript version to fix bugs in pdf compression
+* Prevent PDF rendering Vue warning about tr children of table tags
 * UI: Add a button to expand collapsed menus
 * UI: Fix sections not sorted in sidebar